Move applyMacro from filter to globals
